Create README.md
Bitcoin is a form of digital currency whose aim is to do away with need for central authorities such as banks.Bitcoin makes use of blockchain technology that supports peer-to-peer transactions between its users through a decentralized network.

Bitcoin transactions are authenticated through the proof-of-work consensus mechanism, which gives rewards to cryptocurrency miners for validating its transactions.

Bitcoin was launched in 2009 by a anonymous developer known as Satoshi Nakamoto. It was the first, and has remained as the most valuable asset among cryptocurrencies. Bitcoin is a digital asset that can either be stored in a digital wallet or at a cryptocurrency exchange. Each individual coin represents the value of Bitcoin’s current price in relation to either the US dollar or other world currencies. It's possible to own partial shares of each coin. The smallest unit of each Bitcoin is called a Satoshi. Each Satoshi is equivalent to a hundred millionth of one Bitcoin, so owning fractional shares of Bitcoin is quite common.

Bitcoin is powered by open-source code known as blockchain, which creates a shared public history of transactions organized into "blocks" that are "chained" together to prevent tampering. The technology creates a permanent record of each transaction, and it provides a way for every Bitcoin user to operate with the same understanding of who owns what.

A Bitcoin wallet contains both a public key and a private key, which work together to allow the owner to initiate and digitally sign transactions. This unlocks the central function of Bitcoin — securely transferring ownership from one user to another. Users on the Bitcoin network verify transactions through a process called mining, which confirms that new transactions are consistent with other transactions completed in the past. This ensures that you can’t spend a Bitcoin you don’t have, or that you have previously spent.